reorder

+++ backend/src/utils/seed.ts
@@ -4,14 +4,13 @@ import bcrypt from "bcrypt";
 
 export const seedDatabase = async () => {
   // Check if data already exists
-  // const userCount = await userModel.countDocuments();
-  // if (userCount > 0) {
-  //   console.log('Database already seeded');
-  //   return;
-  // }
+  const userCount = await userModel.countDocuments();
+  if (userCount > 0) {
+    console.log('Database already seeded');
+    return;
+  }
 
   // Clear all residual data
-  await userModel.deleteMany().exec();
   await auctionModel.deleteMany().exec();
 
   const hashedPassword = await bcrypt.hash('password', 10);
